% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_eems.R
\name{average_eems}
\alias{average_eems}
\title{Average the eems within an EEMlist,}
\usage{
average_eems(eemlist)
}
\arguments{
\item{eemlist}{an eemlist compliant with the staRdom/EEM/eemR framework}
}
\description{
Average all the pixels within an eemlist, producing a per-ex/em pixel average EEM object
}
