% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barify.R
\name{barify}
\alias{barify}
\title{A work in progress function to create mirrored offsets of series data.}
\usage{
barify(Input_Data_df, xdata_column = 1, data_column = 2, yoffset = 0.5)
}
\arguments{
\item{Input_Data_df}{The input data frame. Should comprise of at least 1 index (x axis) and the variable (y axis).}

\item{xdata_column}{column for the series variable (x-axis)}

\item{data_column}{column containing the independent/y-axis variable.}

\item{yoffset}{adjustable offset. Use in conjunction with plotting code.}
}
\description{
# Extracts xdata (an index e.g. depth) and ydata and creates
       mirrored offsets, such that the lines between bars in a ggplot2 geom_bar
       plot will be obscured if set to white.At the moment, columns must match
       the default. yoffset can be changed to suit ydata range. This function
       is very much a WIP.
}
