% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_search_nearest.R
\name{binary_search_nearest}
\alias{binary_search_nearest}
\title{Use data.table to find the nearest value in a vector.}
\usage{
binary_search_nearest(data, value)
}
\arguments{
\item{data}{a vector of numeric values to search within}

\item{value}{to drive binary search. i.e. output will be closest value to this input.}
}
\description{
Using binary search provided by the data table package, find the nearest value.
}
