% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_eem.R
\name{check_eem}
\alias{check_eem}
\title{Visually check individual EEMs within an eemlist.}
\usage{
check_eem(eemlist, contour = TRUE, eem_number = 1, output_dir = NULL)
}
\arguments{
\item{eemlist}{A list of EEMs, compliant with the eemR/staRdom framework.}

\item{contour}{TRUE/FALSE to show contours on plot.}

\item{eem_number}{The number of the eem to plot.}

\item{output_dir}{Directory to send a .png of the plot to. If no export is desired, leave as NULL.}
}
\description{
Display and/or export a single EEM, for inspection/checking purposes.
}
