% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_to_eem.R
\name{data_frame_to_eem}
\alias{data_frame_to_eem}
\title{Attempt to convert a raw EEM data dataframe to an eemR-compliant object of class 'eem'.}
\usage{
data_frame_to_eem(
  x,
  sample = NULL,
  filename = NULL,
  location = NULL,
  ex = "cols"
)
}
\arguments{
\item{x}{The target dataframe.}

\item{sample}{Sample/EEM name.}

\item{filename}{Full file name and path of the EEM.}

\item{location}{Optional; where is the EEM from}

\item{ex}{Which side is the excitation axis on - columns, or rows?}
}
\description{
Takes a data frame (e.g. raw EEM data) and attempts to coerce it into an eem object compliant with the eemR/staRdom framework.
}
