% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_bin_emission.R
\name{eem_bin_emission}
\alias{eem_bin_emission}
\title{Bin adjacent emission increments/pixels in an EEM}
\usage{
eem_bin_emission(eem, pix_bins = 2)
}
\arguments{
\item{eem}{An EEM object.}

\item{pix_bins}{number of adjacent emission increments (pixels on a CCD fluoromter) to group together.}
}
\description{
Average adjacent emission pixels in an EEM on a per-excitation wavelength basis. Useful for denoising in high-resolution scans.
}
