% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_data_functions.R
\name{eem_neg_to_0}
\alias{eem_neg_to_0}
\title{Set all negative values within a group of EEMs to 0.}
\usage{
eem_neg_to_0(eemlist, outputfolder = NULL)
}
\arguments{
\item{eemlist}{A list of EEMs in a format compliant with eemR/staRdom.}

\item{outputfolder}{Path to the folder where the new eemlist will be sent. Or NULL, if no export is desired.}
}
\description{
Negative fluorescence is not possible, and typically indicates
       bad processing, noise, or artefacts. This function will set all cases
       of negative fluorescence to 0.
}
