% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_range_mod.R
\name{eem_range_mod}
\alias{eem_range_mod}
\title{Change wavelength range of eems with in an eemlist.}
\usage{
eem_range_mod(eemlist, ex_range, em_range)
}
\arguments{
\item{eemlist}{A list of EEMs compliant with the eemR/staRdom framework.}

\item{ex_range}{Numeric excitation min:max of the new range to constrict the EEMs to.}

\item{em_range}{Numeric excitation min:max of the new range to constrict the EEMs to.}
}
\description{
A tweaked/re-built version of staRdom::eem_range(), which was not
     working for some datasets. Use this function to change the wavelength range
     of a group of EEMs.
}
