% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{eem_read_aqualog_PEM}
\alias{eem_read_aqualog_PEM}
\title{An import function for use with processed excitation-emission matrix (PEM) ASCII files from the HORIBA Aqualog. Essentially identical to eemR::eem_import_aqualog, save for a single line.}
\usage{
eem_read_aqualog_PEM(file)
}
\arguments{
\item{file}{path to file}
}
\description{
To be used within eemR::eem_read() to import aqualog PEM ASCII data files.
}
