% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_data_functions.R
\name{eem_read_mod}
\alias{eem_read_mod}
\title{Read multiple eem files using eemR::eem_read. Essentially identical to that function, but with handling for multiple files.}
\usage{
eem_read_mod(file, recursive = FALSE, import_function)
}
\arguments{
\item{file}{One or more file names.}

\item{recursive}{logical. Should the listing recurse into directories?}

\item{import_function}{Either a character string or a user-defined function to
import a single eem. If a character, it should be one of "cary", "aqualog",
"shimadzu", "fluoromax4".}
}
\description{
Read multiple eem files using eemR::eem_read. Essentially identical to that function, but with handling for multiple files.
}
