% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eemlist_bin_emission.R
\name{eemlist_bin_emission}
\alias{eemlist_bin_emission}
\title{Bin adjacent emission increments/pixels within all EEMs in an eemlist}
\usage{
eemlist_bin_emission(eemlist, pix_bins = 2, verbose = TRUE)
}
\arguments{
\item{eemlist}{A list of EEMs.}

\item{pix_bins}{number of adjacent emission increments (pixels on a CCD fluoromter) to group together.}

\item{verbose}{logical; TRUE/FALSE to print progress. This can be nice as this is not the most efficient set of nested functions - it can take a little while!}
}
\description{
Average adjacent emission pixels in EEMs on a per-excitation wavelength basis. Useful for denoising in high-resolution scans.
}
