% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eemlist_fmax_values.R
\name{eemlist_fmax_values}
\alias{eemlist_fmax_values}
\title{Obtain the fmax value for each EEM in an eemlist.}
\usage{
eemlist_fmax_values(eemlist)
}
\arguments{
\item{eemlist}{a list of eem objects compliant with the staRdom/eemR framework.}
}
\description{
obtain the maximum flurescence intensity value (fmax) from a set of EEMs.
}
