% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eemlist_sd.R
\name{eemlist_sd}
\alias{eemlist_sd}
\title{Generate standard deviations of an eemlist}
\usage{
eemlist_sd(eemlist, mult = 1)
}
\arguments{
\item{eemlist}{A list of EEMs in a format compliant with eemR/staRdom.}

\item{mult}{Numeric; a simple numeric multiplier applied to the resulting ex-em pair sd values.}
}
\description{
Produces an EEM object containing ex/em-pair standard deviations for all eems within the supplied eemlist.
}
