% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eemlist_sp_denoise.R
\name{eemlist_sp_denoise}
\alias{eemlist_sp_denoise}
\title{A single-point scatter denoiser for EEM measurements.}
\usage{
eemlist_sp_denoise(eemlist, interp = TRUE)
}
\arguments{
\item{eemlist}{A list of EEMs in a format compliant with eemR/staRdom.}

\item{interp}{TRUE/FALSE to replace outliers with linear interpolated values instead of NA.}
}
\description{
EEMs are often punctuated by isolated, single points of scatter. These are
     potentially caused by a wide range of factors, including gamma-ray spikes. They can
     severely hamper normalisation attempts, and throw off method detection limit calculations.
     This function provides a simple denoising solution using the tsclean function from the
     forecast package. Data supplied to this function should not contain missing values.
}
