% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eemlist_subtract.R
\name{eemlist_subtract}
\alias{eemlist_subtract}
\title{Subtract an eem from an eemlist}
\usage{
eemlist_subtract(eems_minuend, eem_subtrahend)
}
\arguments{
\item{eems_minuend}{an eemlist compliant with the staRdom/EEM/eemR framework}

\item{eem_subtrahend}{a single EEM.}
}
\description{
Take an eem and subtract it from an eemlist. Useful for, e.g., bulk analytical blank subtraction.
}
