% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eemlist_sum.R
\name{eemlist_sum}
\alias{eemlist_sum}
\title{Sum eems within an eemlist together.}
\usage{
eemlist_sum(eemlist)
}
\arguments{
\item{eemlist}{A list of EEMs in a format compliant with eemR/staRdom.}
}
\description{
Simple addition of the EEM objects within a supplied eemlist.
}
