% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_plotting_functions.R
\name{eempf_comps3D_revex}
\alias{eempf_comps3D_revex}
\title{Plot PARAFAC component modeled EEM in 3D, with a reversed excitation axis.}
\usage{
eempf_comps3D_revex(pfmodel, which = NULL)
}
\arguments{
\item{pfmodel}{A PARAFAC model object.}

\item{which}{NULL or numeric for which component from the PARAFAC model to plot.}
}
\description{
A tweaked staRdom::eempf_comps3D() function that accounts for
     the excitation axis being incorrectly plotted in the original. Seems to
     be a common feature for Aqualog .DAT or .csv originating data. Not quite
     sure where the flipped axis enters the equation!
}
