% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emscan_read.R
\name{emscan_read}
\alias{emscan_read}
\title{An importer for Aqualog-produced 2D emission scans}
\usage{
emscan_read(files, type = "sample")
}
\arguments{
\item{files}{character vector containing one or more full filenames of emission scans}

\item{type}{character - one of either 'sample' or 'blank'}
}
\description{
A simple importing function for 2D emission scans produced using the
     HORIBA Aqualog. Supports both sample and blank scan types (though not both at the same time)
}
