% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ramanpeak_areas.R
\name{extract_ramanpeak_areas}
\alias{extract_ramanpeak_areas}
\title{Extract the peak areas from spectrometer Raman curves, for EEM normalisation purposes.}
\usage{
extract_ramanpeak_areas(
  RAMdat,
  range_lower = NULL,
  range_upper = 500,
  method = "aqualog",
  halfwidth = 1800,
  sequence = 8,
  tolerance = 0.05,
  landa = 350,
  output_dir = NULL,
  gif = FALSE
)
}
\arguments{
\item{RAMdat}{A data frame containing wavelength increments in column 1, and fluorescence intensities in subsequent columns.}

\item{range_lower}{optional; a lower limit for calculations.}

\item{range_upper}{optional; an upper limit for calculations. 500 by default to avoid higher-order scattering bands.}

\item{method}{one of 'aqualog' or 'RIR' for either default Aqualog 380:410nm, or to use RamanIntegrationRange's gradient calculations to determine peak position, and thus area.}

\item{halfwidth}{if method = 'RIR', what is the halfwidth? 1800 by default.}

\item{sequence}{if method = 'RIR', what is the sequence? 8 by default.}

\item{tolerance}{if method = 'RIR', what is the gradient tolerance? 0.05 by default.}

\item{landa}{what is the excitation band of the raman spectra? 350 (nm) by default.}

\item{output_dir}{optional; path to the output directory of peak images.}

\item{gif}{optional; create .gif animation of peak images?}
}
\description{
This function provides two methods to calculate the area under the
      Raman peak of water for fluorescence data normalisation purposes. The first
      method is a port of RamanIntegrationRange from the MATLAB drEEM package,
      which uses a gradient detection method to identify the start and end of
      the Raman peak of water. The second is a standard Aqualog method, which
      simply assumes that the peak extends from 380 to 410nm. Plot export and
      .gif wrapping are optional. Some example raman curves are included in
      the eemUtils data folder.
}
