% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrpf_loadings.R
\name{extrpf_loadings}
\alias{extrpf_loadings}
\title{Extract fluorescence intensity loadings from a PARAFAC model.}
\usage{
extrpf_loadings(pfmodel, by_index = FALSE, fill_stat = FALSE)
}
\arguments{
\item{pfmodel}{A PARAFAC model object containing any number of components}

\item{by_index}{TRUE/FALSE index the loadings by the sample names via registry matching. Requires numerals in the sample names}
}
\description{
Extract the per-sample modeled intensity loadings for each component
      from a PARAFAC models. Optional numeric indexing of sample names
      using regmatching, assuming there are numerals to extract. Use lapply for
      multiple models.
}
