% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrpf_loadings_percent.R
\name{extrpf_loadings_percent}
\alias{extrpf_loadings_percent}
\title{Extract the per-EEM percentage contributions of each component}
\usage{
extrpf_loadings_percent(pfmodel, eemlist, denormalise = FALSE)
}
\arguments{
\item{pfmodel}{a PARAFAC model object}

\item{eemlist}{supply the eemlist used for denormalisation. Only necessary if denormalise is set to TRUE.}

\item{denormalise}{TRUE/FALSE to denormalise the loadings prior to the percent calculation}
}
\description{
Return the percent contribution of each component to the modelled fluorescence response of each sample, using
     the modelled loadings
}
