% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrpf_peak_positions.R
\name{extrpf_peak_positions}
\alias{extrpf_peak_positions}
\title{Return the ex/em position of the peak maxima for a PARAFAC modeled component.}
\usage{
extrpf_peak_positions(pfmodel, component = NULL)
}
\arguments{
\item{pfmodel}{A PARAFAC model object.}

\item{component}{NULL for all components, or the numeric identifier of a single component.}
}
\description{
Uses extrpf_peak_spectra to obtain the single ex/em coordinate pair for the maxima of the
    specified PARAFAC component.
}
