% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrpf_peak_spectra.R
\name{extrpf_peak_spectra}
\alias{extrpf_peak_spectra}
\title{Extract spectra at the peak wavelength position.}
\usage{
extrpf_peak_spectra(pfmodel, component = 1)
}
\arguments{
\item{pfmodel}{A PARAFAC model object}

\item{component}{NULL or numeric for specific components.}
}
\description{
Extract the spectra at the peak for a given component.
}
