% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrpf_residual_eems.R
\name{extrpf_residual_eems}
\alias{extrpf_residual_eems}
\title{Extract per-sample residual EEM objects}
\usage{
extrpf_residual_eems(pfmodel, eemlist)
}
\arguments{
\item{pfmodel}{a single PARAFAC model object returned from staRdom::eem_parafac containing any number of components.}

\item{eemlist}{The eemlist object passed to the above PARAFAC model.}
}
\description{
Extract the 'leftover' fluorescence (i.e. residuals) in each EEM passed to a given PARAFAC model. Areas of
    high (negative) fluorescence indicate regions where the PARAFAC model is under (over) fitted.
}
