% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrpf_residuals.R
\name{extrpf_residuals}
\alias{extrpf_residuals}
\title{Calculate and extract PARAFAC model residuals}
\usage{
extrpf_residuals(
  pfmodel,
  eem_list,
  select = NULL,
  cores = parallel::detectCores(logical = FALSE) - 1,
  denormalise = FALSE,
  extend_eemlist = TRUE,
  verbose = FALSE,
  force_names = TRUE
)
}
\arguments{
\item{pfmodel}{a PARAFAC model object from staRdom::eem_parafaC()}

\item{eem_list}{a list of eem objects compliant with the staRdom/eemR framework}

\item{select}{a character vector containing the names of the desired samples. See ?staRdom::eem_parafac}

\item{cores}{number of cores used for multi-threading. See ?staRdom::eem_parafac}

\item{denormalise}{is the PARAFAC model normalised? If so, use the eemlist to denormalise the loadings. Optional; only effects relative magnitude of residuals}

\item{extend_eemlist}{TRUE/FALSE to ensure equal EEM sizes before calculations. Uses staRdom::eem_extend2largest if differences are found.}

\item{verbose}{return various messages during operation.}
}
\description{
An expansion upon staRdom::eempf_residuals. Minor internal code changes to fix a couple of bugs, and extra options added.
}
