% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_data_functions.R
\name{extrpf_spectra_or_eems}
\alias{extrpf_spectra_or_eems}
\title{Extract modeled spectra or EEM objects from a pfmodel list.}
\usage{
extrpf_spectra_or_eems(pfmodel_list, type = 1)
}
\arguments{
\item{pfmodel_list}{PARAFAC model list.}

\item{type}{numeric 1 or 2 - extract EEMs or peaks, respectively.}
}
\description{
Extracts either the component peak spectra, or the modelled EEMs,
    from a pfmodel list. Essentially produces the object used to plot the two
    types of staRdom::eempf_plot_comps().
}
