% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_plotting_functions.R
\name{fmax_corrplot}
\alias{fmax_corrplot}
\title{Plot PARAFAC model loadings versus per-sample Fmax at the component peak position.}
\usage{
fmax_corrplot(
  pfmodel,
  eemlist,
  component = 1,
  denorm = TRUE,
  labels = FALSE,
  label_threshold = 2
)
}
\arguments{
\item{pfmodel}{A PARAFAC model object.}

\item{eemlist}{An eemlist object.}

\item{component}{Integer. Which PARAFAC component to plot?}

\item{denorm}{TRUE/FALSE denormalise the loadings based upon eemlist (not parafac) fmax values?}

\item{labels}{TRUE/FALSE extract numeric values from sample names and use these as labels. values/10 by default to match a lexicographic scheme.}

\item{label_threshold}{Integer. Values lying above this number multiplied by the mean will be given labels via geom_text}
}
\description{
A simple scatter plot of PARAFAC component loadings versus Fmax data. Optional labelling,
     derived through extraction of numeric values in the sample names.
}
