% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_data_functions.R
\name{get_pf_peak_spectra}
\alias{get_pf_peak_spectra}
\title{Extract spectra at the peak wavelength position.}
\usage{
get_pf_peak_spectra(pfmodel, component = 1)
}
\arguments{
\item{pfmodel}{A PARAFAC model object a la staRdom.}

\item{component}{Extract spectra for this component. Numeric.}
}
\description{
An alternative to the extrpf_peaks_or_eems. Simply extract the spectra at the peak for a given component.
}
