% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_data_functions.R
\name{get_pfload_percent}
\alias{get_pfload_percent}
\title{Derive the per-component percentage intensity contribution}
\usage{
get_pfload_percent(loadings)
}
\arguments{
\item{loadings}{An output data frame from extrpf_loadings()}
}
\description{
Determine the per-sample percentage contribution of each PARAFAC component
      across a given set of sample PARAFAC intensity loadings.
}
