% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_textsize_scale.R
\name{ggplot_textsize_scale}
\alias{ggplot_textsize_scale}
\title{A simple textsize scaler for ggplot objects.}
\usage{
ggplot_textsize_scale(
  plot,
  multiplier = 0.8,
  default_title_size = 11,
  default_axis_title_size = 10,
  default_axis_text_size = 9,
  default_legend_title_size = 10,
  default_legend_text_size = 9
)
}
\arguments{
\item{plot}{a ggplot2 object.}

\item{multiplier}{numeric; a direct multiplier for specified text sizes.}

\item{default_title_size}{numeric; default text size for plot title.}

\item{default_axis_title_size}{numeric; default text size for axis titles.}

\item{default_axis_text_size}{numeric; default text size for axis text.}

\item{default_legend_title_size}{numeric; default text size for legend titles.}

\item{default_legend_text_size}{numeric; default text size for legend text}
}
\description{
Takes a set of text size alteration numbers and modifies a ggplot. Useful
     when attempting to get a ggplot object's text scaling right.
}
