% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_row.R
\name{insert_row}
\alias{insert_row}
\title{Insert a single row seamlessly into a data frame.}
\usage{
insert_row(dataframe, newrow_index, newrow_contents)
}
\arguments{
\item{dataframe}{The target data frame.}

\item{newrow_index}{numeric - row position will the new row occupy?}

\item{newrow_contents}{The new row to be inserted into the dataframe.}
}
\description{
Insert a row into a data frame whilst shifting the other rows to
     make room.
}
