% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_eemlist.R
\name{normalise_eemlist}
\alias{normalise_eemlist}
\title{Normalise EEM data using one of a number of different methods.}
\usage{
normalise_eemlist(
  eemlist,
  type = "max_eem_val",
  norm_log,
  norm_data,
  norm_value = NULL,
  outputfolder = NULL,
  append_name = NULL
)
}
\arguments{
\item{eemlist}{A list of EEMs, compliant with the eemR/staRdom framework.}

\item{type}{one of 'raman_peak_to_area','standard_from_list','standard_single_value', or 'max_eem_val'.}

\item{norm_log}{A dataframe listing all eems in the eemlist, along with which sample group they belong to.}

\item{norm_data}{A dataframe listing the normalisation values (old and new) corresponding to each sample group.}

\item{norm_value}{If a single value is being used, a numeric value. If not, leave as NULL.}

\item{append_name}{Extension to be added to the name of each EEM.}

\item{output_folder}{Path for export of normalised EEMs.}
}
\description{
Normalise EEMs in a dataset using different methods: convert from
     raman peak to raman area (Raman Units), using values supplied from a list,
     using a single value, or internally normalise each EEM by its own maximum intensity.
}
