% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfluor_parse.R
\name{openfluor_parse}
\alias{openfluor_parse}
\title{Import and parse a downloaded export .csv file from the OpenFluor databse}
\usage{
openfluor_parse(file, normalise_spectra = TRUE)
}
\arguments{
\item{file}{full filepath to the .csv file.}

\item{normalise_spectra}{TRUE/FALSE; logical to normalise excitation and emission spectra from the database (irrelevant/automatic during the TCC comparisons but can help with plotting)}
}
\description{
The OpenFluor database allows for the statistical comparison of a PARAFAC model with published spectra.
  These comparisons can be downloaded as a .csv file, which this function can parse into a sorted list object containing all
  the information within the file. It is recommended to first open and immediately save the downloaded .csv in e.g. MS Excel as the
  raw file is slightly finnicky with read.csv and read_csv.
}
