% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_data_functions.R
\name{peak_positions}
\alias{peak_positions}
\title{Determines the peak positions for a PARAFAC modeled component.}
\usage{
peak_positions(modeled_spectra)
}
\arguments{
\item{modeled_spectra}{PARAFAC-modeled spectra. An output from extrpf_spectra_or_eems(type = 2)}
}
\description{
DEFUNCT Determine the Ex/Em peak position of a PARAFAC component. Takes
     an output from extrpf_spectra_or_eems(type = 2).
}
