% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extrpf_eems.R
\name{plot_extrpf_eems}
\alias{plot_extrpf_eems}
\title{Plot PARAFAC component modeled EEMs.}
\usage{
plot_extrpf_eems(modeled_EEMs, contour = FALSE)
}
\arguments{
\item{modeled_EEMs}{An output from extrpf_spectra_or_eems(type = 1).}

\item{contour}{TRUE/FALSE to display contours on the EEM plot/s.}
}
\description{
Plot the modeled EEMs from one or more PARAFAC models. Takes an
     output from extrpf_spectra_or_eems(type = 1).
}
