% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_plotting_functions.R
\name{plot_extrpf_spectra}
\alias{plot_extrpf_spectra}
\title{Plot PARAFAC component peak maxima spectra.}
\usage{
plot_extrpf_spectra(modeled_spectra, eemlist, title)
}
\arguments{
\item{modeled_spectra}{An output from extrpf_spectra_or_eems(type = 2).}

\item{eemlist}{The list of EEMs used to produce the PARAFAC.}

\item{title}{A character vector, to be displayed as a title on the plot.}
}
\description{
Plot the modeled spectra from one or more PARAFAC models. Takes
     an output from extrpf_spectra_or_eems(type = 2).
}
