% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_plotting_functions.R
\name{plot_pfmodels_spectra}
\alias{plot_pfmodels_spectra}
\title{Plot the component spectra intersecting at component maxima from a list of PARAFAC models.}
\usage{
plot_pfmodels_spectra(pfmodel_list, eemlist, title)
}
\arguments{
\item{pfmodel_list}{PARAFAC model list.}

\item{eemlist}{list of EEMs used to produce the pfmodels.}

\item{title}{A title to add to the plot.}
}
\description{
This is an unpacked staRdom::eempf_plot_comps(type = 2).
}
