% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_eemlist_csvs.R
\name{save_eemlist_csvs}
\alias{save_eemlist_csvs}
\title{Save a list of EEMs as individual .csv files.}
\usage{
save_eemlist_csvs(eemlist, outputfolder = NULL, append_name = NULL)
}
\arguments{
\item{eemlist}{A list of EEMs in a format compliant with eemR/staRdom.}

\item{outputfolder}{Either the path of the folder to write the .csvs to, or NULL for the working directory.}

\item{append_name}{optional; a character vector that will be appended to each .csv file.}
}
\description{
# Save the EEMs comprising an eemlist to a set of csv files. If no
     output_dir is set, the csvs are written straight to the working directory.
}
