% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_data_functions.R
\name{A_missing_mod}
\alias{A_missing_mod}
\title{Calculate the sample loadings for samples not involved in model building}
\usage{
A_missing_mod(
  eem_list,
  pfmodel = NULL,
  skip_reduce = TRUE,
  cores = parallel::detectCores(logical = FALSE),
  components = NULL,
  const = NULL,
  control = NULL,
  ...
)
}
\arguments{
\item{eem_list}{object of class eemlist with sample data}

\item{pfmodel}{object of class parafac}

\item{skip_reduce}{TRUE/FALSE to force reduction of EEM dimensions to their smallest member. Required for modelling, but can cause errors in practice.}

\item{cores}{number of cores to use for parallel processing}

\item{components}{optionally supply components to use manually, either as a variable of class parafac_components or as a list of variables of class parafac_components, if you do so,}

\item{const}{optional constraints for model, just used, when components are supplied}

\item{control}{optional constraint control parameters for model, just used, when components are supplied}

\item{...}{additional arguments passed to eem_parafac}
}
\description{
This is simply a port of staRdom's A_missing function, with an option to skip reduction of EEMs.
}
