% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_emscan.R
\name{bin_emscan}
\alias{bin_emscan}
\title{Bin adjacent emission increments/pixels.}
\usage{
bin_emscan(em_spectra, pix_bins = 2, average = TRUE)
}
\arguments{
\item{em_spectra}{a long-form (gathered) emission spectra.}

\item{pix_bins}{number of adjacent emission increments (pixels on a CCD fluoromter) to group together.}
}
\description{
Average adjacent emission pixels to reduce noise in an emission spectra.
}
