% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_data_functions.R
\name{eem_bin}
\alias{eem_bin}
\title{Bin the intensity values from an EEM.}
\usage{
eem_bin(eem, nbins = 12)
}
\arguments{
\item{eem}{An eem object, compliant with the eemR/staRdom packages.}

\item{nbins}{The number of bins. 12 by default.}
}
\description{
Takes the intensity values of an EEM and bins them based upon a preset number of
       bins. The lowest break-end is the smallest intensity value, and the highest will be Inf, such
       that the top bin is
       (max intensity - ((max intensity)/nbins)):Inf
}
