% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_data_functions.R
\name{eemdf_to_eem}
\alias{eemdf_to_eem}
\title{Convert a dataframe to an EEM object of the style used by EEM/eemR/staRdom.}
\usage{
eemdf_to_eem(eemdf, file, sample, location, gathered = FALSE)
}
\arguments{
\item{eemdf}{the dataframe to be coerced to an EEM object.}

\item{file}{filename of the EEM, if applicable.}

\item{sample}{the samplename of the EEM, if applicable.}

\item{location}{the location of the EEM file, if applicable.}

\item{gathered}{TRUE/FALSE is the eemdf in a short (not gathered; FALSE) or a long (gathered; TRUE) format?}
}
\description{
An alternative to eemR's eem constructor. Intended to reverse as.data.frame(eem).
}
