% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_data_functions.R
\name{eemlist_average}
\alias{eemlist_average}
\title{Average a set of EEMs.}
\usage{
eemlist_average(eemlist)
}
\arguments{
\item{eemlist}{A list of EEMs, compliant with the eemR/staRdom framework.}
}
\description{
Average the EEMs within an eemlist.
}
