% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_data_functions.R
\name{eemlist_sd}
\alias{eemlist_sd}
\title{Generate standard deviations of an eemlist}
\usage{
eemlist_sd(eemlist, mult = 1)
}
\arguments{
\item{eemlist}{A list of EEMs in a format compliant with eemR/staRdom.}

\item{mult}{Numeric; a simple numeric multiplier applied to the resulting matrix.}
}
\description{
Produces element-wise sd data for all eems with the supplied eemlist.
}
