% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_plotting_functions.R
\name{extract_procstep_eems}
\alias{extract_procstep_eems}
\title{Extract a specific EEM from a set of eemlists}
\usage{
extract_procstep_eems(
  list_of_eemlists,
  which_eem = 1,
  output_dir = NULL,
  contour = TRUE
)
}
\arguments{
\item{which_eem}{integer - numerically index which eem within the eemlists?}

\item{output_dir}{NULL or full path to destination directory in which to save a .png image of the EEMs together.}

\item{contour}{TRUE/FALSE include contours in exported plots, assuming output_dir != NULL}

\item{list_of_eemlist}{a list vector, with each element comprising an eemlist compliant with the staRdom/eemR framework}
}
\description{
when given a list of eemlists representing different stages in a processing
   framework, extract a single EEM from each eemlist in order to show the processing steps.
}
