% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_data_functions.R
\name{extrpf_eems}
\alias{extrpf_eems}
\title{Extract the modelled EEMs from a given PARAFAC model as an eemlist.}
\usage{
extrpf_eems(pfmodel)
}
\arguments{
\item{pfmodel}{a PARAFAC model object containing one or more components}
}
\description{
Pulls out the modelled EEMs from a PARAFAC model, in the standard eemlist format,
     comprised of eem objects compliant with the staRdom/EEM/eemR framework.
}
