% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_data_functions.R
\name{extrpf_fmax}
\alias{extrpf_fmax}
\title{Derive component Fmax values for a given PARAFAC model and eemlist.}
\usage{
extrpf_fmax(pfmodel, eemlist)
}
\arguments{
\item{pfmodel}{A single PARAFAC model object containing any number of components}

\item{eemlist}{a list of eems in the staRdom/eemR compliant format}
}
\description{
Fluorescence intensity at the component maxima (peak Ex/Em wavelength) can be used to infer intensity.
       Assuming a well-performing PARAFAC model, these values should correlate extremely well with the component
       A-mode values/loadings.
}
