% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrpf_loadings.R
\name{extrpf_loadings}
\alias{extrpf_loadings}
\title{Extract fluorescence intensity loadings from a list of PARAFAC models.}
\usage{
extrpf_loadings(pfmodel_list, by_index = FALSE)
}
\arguments{
\item{pfmodel_list}{A list of PARAFAC models, typically an output from stardom::eem_parafac()}

\item{by_index}{TRUE/FALSE index the loadings by the sample names via registry matching. Requires numerals in the sample names.}
}
\description{
Extract the per-sample modeled intensity loadings for each component
      from a list of PARAFAC models. Optional numeric indexing of sample names
      using regmatching, assuming there are numerals to extract.
}
