% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_data_functions.R
\name{extrpf_loadings_denorm}
\alias{extrpf_loadings_denorm}
\title{Multiply the loadings values of PARAFAC components by normalisation factors}
\usage{
extrpf_loadings_denorm(pfmodel, eemlist, type = "short")
}
\arguments{
\item{pfmodel}{a PARAFAC model object, typically an individual ouptut from staRdom::eem_parafaC()}

\item{eemlist}{a list of eem objects compliant with the staRdom/eemR framework}

\item{type}{short or long format. Short by default. Long format data works better for grouping in ggplot}
}
\description{
A PARAFAC model fed with normalised eems outputs by default loadings that are less useful for direct
      sample-to-sample comparison. This can be remedied by multiplying the loadings by each eem fmax value (i.e. the
      normalisation factors used to normalise the eems originally).
}
