% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_data_functions.R
\name{extrpf_peak_positions}
\alias{extrpf_peak_positions}
\title{Return the ex/em position of the peak maxima for a PARAFAC modeled component.}
\usage{
extrpf_peak_positions(pfmodel, component = 1)
}
\arguments{
\item{modeled_spectra}{PARAFAC-modeled spectra. An output from extrpf_spectra_or_eems(type = 2)}
}
\description{
Uses extrpf_peak_spectra to obtain the single ex/em coordinate pair for the maxima of the
    specified PARAFAC component.
}
