% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_data_functions.R
\name{Generate_CORCONDIA}
\alias{Generate_CORCONDIA}
\title{Return a table of Core Consistency Diagnostic values for a given group of PARAFAC models}
\usage{
Generate_CORCONDIA(pfmodel, eemlist)
}
\arguments{
\item{pfmodel}{A group of PARAFAC model objects, typically outputs from staRdom::eem_parafac()}

\item{eemlist}{The group of EEMs used to run the above PARAFAC model}
}
\description{
A simple wrapper for stardom::eempf_corcondia().
}
