% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_data_functions.R
\name{get_eem_max_coords}
\alias{get_eem_max_coords}
\title{Get the max intensity value's ex/em position.}
\usage{
get_eem_max_coords(eem, verbose = FALSE)
}
\arguments{
\item{eem}{an eem object compliant with the staRdom/eemR framework}

\item{verbose}{TRUE/FALSE to return a message with the identified max position.}
}
\description{
Get the excitation/emission coordinates of the point of maximum fluorescence
     intensity within the target EEM.
}
