% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_plotting_functions.R
\name{ggeem2}
\alias{ggeem2}
\title{A tweaked EEM plotter, building off of ggeem() from staRdom()}
\usage{
ggeem2(
  eem,
  fill_max = FALSE,
  bin_vals = 12,
  colpal = "12pal",
  contour = TRUE,
  interpolate = FALSE,
  redneg = NULL,
  legend = TRUE,
  textsize_multiplier = 1,
  ...
)
}
\arguments{
\item{eem}{An eem object compliant with the staRdom/eemR framework}

\item{bin_vals}{NULL or numeric for number of bins. If no binning is desired, set to NULL. Defaults to 12, the length of the default colour palette, eem_palette_12.}

\item{colpal}{Provide a colour palette as a vector of colours. Two defaults are available - "12pal" for a 12-colour palette based on MATLAB's jet scheme, or "rainbow" for ggeem's default scheme.}

\item{contour}{TRUE/FALSE to plot contour}

\item{interpolate}{TRUE/FALSE to interpolate. Fairly certain this is defunct; refer to ?staRdom::ggeem()}

\item{redneg}{refer to ?staRdom::ggeem()}

\item{legend}{TRUE/FALSE to display legend}

\item{textsize_multiplier}{a simple numeric multiplier for increasing text size of graphical elements. To help with R's tricky graphics device text scaling when exporting.}
}
\description{
An update to staRdom's existing EEM plotter, ggeem. Option for binning values,
     along with new colours amongst other things.
}
