% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_data_functions.R
\name{interpolate_eem}
\alias{interpolate_eem}
\title{Add data points between existing values across a specified EEM axis.}
\usage{
interpolate_eem(eem, n_pp = 2, direction = "ex", verbose = FALSE)
}
\arguments{
\item{eem}{An eem object compliant with the staRdom/eemR framework.}

\item{n_pp}{numeric; how many data points to add between existing ones?}

\item{Direction}{"ex" only.}
}
\description{
Interpolate between existing eem data points, adding values across a given axis.
    Intended for data to be plotted, as an alternative to staRdom's eem_smooth() smoothing function.
    Support only for excitation axis at this stage. Not to be confused with staRdom's eem_interp.
}
