% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_data_functions.R
\name{knmi_monthly_rescale}
\alias{knmi_monthly_rescale}
\title{Work in progress function to bin or interpolate series data at a monthly resolution.}
\usage{
knmi_monthly_rescale(
  data,
  old_timescale,
  new_timescale,
  n_month_groups = 12,
  method = "bin",
  new_timescale_centered = FALSE,
  replace_NA = TRUE,
  annual_average_startmonth = NULL,
  average_type = "impersonate_monthly",
  average_alignment = NULL,
  align_monthly_impersonation = FALSE
)
}
\arguments{
\item{data}{The sub-annually resolved series data to be re-scaled.}

\item{old_timescale}{the original time index for the data.}

\item{new_timescale}{the new timescale to be re-scaled into. E.g. seq(2010,2000,1/12)}

\item{n_month_groups}{placeholder - number of months. Leave as 12, as that's how many months there are I guess.}

\item{method}{'bin' or 'interpolate' - how will the rescaling be performed? Binning is usually better.}

\item{new_timescale_centered}{TRUE/FALSE if binning, will the bins be centered?}

\item{replace_NA}{TRUE/FALSE to replace instances of NA with -999.9, which is KNMI's default missing value handle.}

\item{annual_average_startmonth}{NULL or numeric between 1 and 12. If set to a number between 1 and 12, an annual average will be calculated starting at that month.}

\item{average_type}{one of either "annual_averages" or "impersonate_monthly".. The former will return the average values in data frame, and the latter will impersonate a format that KNMI will accept as a monthly resolution data to allow comparison of the annual averages against monthly-resolved data.}

\item{average_alignment}{NULL or one of either "current" or "previous". NULL means that Jan to June will (months 1 to 6) will commence the average in the current year, and July to Dec in the preceding year.}

\item{align_monthly_impersonation}{TRUE/FALSE. If returning an impersonate_monthly data frame (see previous param), should the annual values start in January for a given year, or commence at the annual_average_startmonth?}
}
\description{
Bin or interpolate data to a monthly scale and output in a dataframe
       that is compatible with the KNMI climate explorer user upload field.
       Originally developed for use with sub-annually resolved firn core fluorescence
       intensity data.
}
