% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_plotting_functions.R
\name{plot_comp_spectra}
\alias{plot_comp_spectra}
\title{Plot PARAFAC model Ex and Em mode (B and C) loadings}
\usage{
plot_comp_spectra(pfmodel, comp)
}
\arguments{
\item{pfmodel}{A PARAFAC model object.}

\item{eemlist}{An eemlist object.}

\item{component}{Integer. Which PARAFAC component to plot?}

\item{denorm}{TRUE/FALSE denormalise the loadings based upon eemlist (not parafac) fmax values?}

\item{labels}{TRUE/FALSE extract numeric values from sample names and use these as labels. values/10 by default to match a lexicographic scheme.}

\item{label_threshold}{Integer. Values lying above this number multiplied by the mean will be given labels via geom_text}
}
\description{
Basic lineplot of Ex and Em spectra extracted from PARAFAC model components.
}
