% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_plotting_functions.R
\name{plot_eem_spectra}
\alias{plot_eem_spectra}
\title{Plot spectra from an EEM at a supplied Ex/Em position}
\usage{
plot_eem_spectra(eem, ex, em)
}
\arguments{
\item{eem}{a single EEM object, compliant with the staRdom/eemR framework}

\item{ex}{numeric excitation wavelength value}

\item{em}{numeric emission wavelength value}
}
\description{
A simple ggplot2 lineplot producer, using slice_eem to extract Ex and Em
     spectra at the given coordinates.
}
