% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_plotting_functions.R
\name{save_eemlist_pngs}
\alias{save_eemlist_pngs}
\title{Save all the EEMs within an eemlist object as .png images}
\usage{
save_eemlist_pngs(eemlist, contour = FALSE, output_dir = NULL)
}
\arguments{
\item{eemlist}{the target eemlist object compliant with the staRdom/eemR framework.}

\item{contour}{TRUE/FALSE to add contour to EEM.}

\item{outpud_dir}{full path to the destination folder.}
}
\description{
loop through an eemlist, saving EEM data as individual .png images.
}
