% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_data_functions.R
\name{slice_eem}
\alias{slice_eem}
\title{Extract spectra from an EEM at a given Ex/Em wavelength.}
\usage{
slice_eem(eem, ex, em)
}
\arguments{
\item{eem}{an eem object compliant with the staRdom/eemR framework}

\item{ex}{an excitation wavelength value, in nm}

\item{em}{an emission wavelength value, in nm}
}
\description{
'Slice' an EEM at a given Ex/Em coordinate pair, and extract the spectra at that location.
}
